reversetelleport

+++ Assets/Scripts/Bullet.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Bullet : MonoBehaviour {
     public int damadge = 1;
-    public int lifeTime;
+    public float lifeTime;
     public float fireRate;
     // Use this for initialization
     public bool isEnemyShoot = false;
